configuration updated to ses2 format. Use includes/configs/*.xml for adding new configs template manager updates from ses2 plus changing in how files are cached...removed minify and used only 2 classes jsmin and cssmin. at first sight it seems faster and better compression removed profiler and added query free objects removed from catalog all references to old datacess class or any other classes we don't use. I tested in some casses but it may give some problems  some cleaning in functions  template manager layout backups  template manager import export with zip file from hard drive to hard drive
credit card exp reminder
 sort order of modules not working
 set all products taxable
 allow deleting more than one product
 line-height problem(may need some testing)
 shipping not updating right when updating quantity or deleting products from cart on checkout also bug when updating quantity
 checkout/success is not more ajax, so you can redirect payment modules to checkout success directly
if installing multi-store have any top level admins get access to all stores
product scroller widget: add a "Auto Scroll" checkbox also solved the problem when scroller was going crazy after some time.
removed social network widget and renamed the addtoany...it may give problems on some sites so you need to add the new socialNetwork widget and configure it accordigly
extra fees and time fees to ppr and also some improvements to order/product shipping selection, zip selection made mandatory if zip codes are used on ppr.

+++ admin/applications/admin_members/actions/deleteGroup.php
@@ -1,8 +1,14 @@
 <?php
-	$AdminGroups = Doctrine_Core::getTable('AdminGroups')->findByAdminGroupsId((int)$_GET['gID']);
+$success = false;
+if ((int)$_GET['gID'] > 1){
+	$AdminGroups = Doctrine_Core::getTable('AdminGroups')->find((int)$_GET['gID']);
 	if ($AdminGroups){
 		$AdminGroups->delete();
+		$success = true;
 	}
+}
 
-	EventManager::attachActionResponse(itw_app_link(null, null, 'groups'), 'redirect');
+	EventManager::attachActionResponse(array(
+		'success' => $success
+	), 'json');
 ?>
